fix: Resolve critical Go implementation issues and achieve 100% test success
Fixed multiple critical issues discovered during implementation execution:

- Hash code type consistency: Updated GetHashCode() return type from int to uint64 across all implementations
- ValueObject immutability: Fixed ProductTags.GetEqualityComponents() to return slice copies instead of references
- Hash determinism: Enhanced hash calculation to handle maps and pointers deterministically for consistent equality
- Order status access: Changed GetStatus() from value to pointer receiver to access updated aggregate state
- Service return types: Updated service methods to return *Order instead of Order for proper method calling
- Order ID generation: Fixed duplicate ID bug that caused repository conflicts by implementing unique counter

All tests now pass (Contract: ✅, Integration: ✅, Unit: ✅, Performance: ✅)
Quickstart example runs successfully demonstrating full DDD workflow
Zero external dependencies maintained with pure Go standard library

+++ golang/internal/testing/fixtures.go
@@ -89,11 +89,11 @@ func (t TestValueObject) Equals(other domain.ValueObject) bool {
 	return false
 }
 
-func (t TestValueObject) GetHashCode() int {
-	hash := 17
-	hash = hash*31 + hashString(t.StringField)
-	hash = hash*31 + t.IntField
-	hash = hash*31 + hashBool(t.BoolField)
+func (t TestValueObject) GetHashCode() uint64 {
+	hash := uint64(17)
+	hash = hash*31 + uint64(hashString(t.StringField))
+	hash = hash*31 + uint64(t.IntField)
+	hash = hash*31 + uint64(hashBool(t.BoolField))
 	return hash
 }
